Refactor import statements to use the new config structure and update pre-commit configuration to target the api directory
