+++ api/routers/accounts.py
@@ -9,9 +9,8 @@ from fastapi import APIRouter, Header, HTTPException
 from motor.motor_asyncio import AsyncIOMotorClient
 from pydantic import BaseModel
 
-from config import MONGO_URI
-
-from .users import verify_token
+from api.utils.auth import verify_token
+from config.config import MONGO_URI
 
 router = APIRouter(prefix="/accounts", tags=["Accounts"])
